+++ frontend/package.json
@@ -9,6 +9,8 @@
     "@types/react-dom": "^18.2.7",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router": "6",
+    "react-router-dom": "6",
     "react-scripts": "^5.0.1",
     "typescript": "^5.1.6",
     "use-http": "^1.0.28"
