+++ frontend/src/pages/About/About.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const About = () => <h4 className="text-md pt-2">This page is the about page</h4>;
+
+export default About;
