+++ frontend/src/pages/PageNotFound/PageNotFound.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const PageNotFound = () => (
+  <h4 className="text-md pt-2 text-rose-600">The page you are looking is not found.</h4>
+);
+
+export default PageNotFound;
